+++ agenda/README.md
@@ -18,7 +18,15 @@ npm run build
 ### Lints and fixes files
 ```
 npm run lint
-```
 
+```
+### Iniciar servidor 
+
+```
+node index_servidor
+
+```
+## Possiveis erros 
+caso de erro na importação de alguns arquivos na base do servidor só modifique ( descomente ) as partes com "import " e comente as parte que vem com "require". 
 ### Customize configuration
 See [Configuration Reference](https://cli.vuejs.org/config/).
