+++ agenda/package.json
@@ -12,6 +12,7 @@
     "bootstrap-vue": "^2.23.1",
     "core-js": "^3.8.3",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.21.0",
     "mysql": "^2.18.1",
     "todolist": "file:",
